+++ src/lib.rs
@@ -65,6 +65,7 @@ pub mod audio;
 pub mod audio_mix;
 pub mod duration;
 pub mod id;
+pub mod light;
 pub mod object;
 pub mod ops;
 pub mod page;
@@ -88,6 +89,7 @@ pub use audio::{AudioCue, AudioSource, DuckBus, Generator};
 pub use audio_mix::mix_cues;
 pub use duration::{Lifetime, SceneDuration, TimeStamp};
 pub use id::ObjectId;
+pub use light::{Light, LightCommon, SpotParams};
 pub use object::{
     BlendMode, Canvas, ClipRect, Effect, ImageSource, LengthUnit, LiveStreamHandle, ObjectKind,
     Sample, SceneObject, Shape, TextRun, Transform, VideoSource,